Fix CRITICAL u64 overflow in claim_payout + distributeReward idempotency guard
On-chain (deployed to devnet):
- claim_payout payout math now uses u128 intermediate multiplication to
  prevent overflow when betting pools exceed ~4 SOL. Without this fix,
  winning bettors could not claim and SOL was permanently locked.

Off-chain:
- Added getRumbleTxSignature() helper to check if a tx step was already done
- distributeReward now checks Supabase for existing signature before calling,
  preventing double-distribution of ICHOR rewards on retry/restart

+++ packages/solana/programs/rumble-engine/src/lib.rs
@@ -1548,12 +1548,14 @@ pub mod rumble_engine {
 
             // Bettor's proportional share of the allocation
             // share = (bettor_winning_deployed / first_pool) * place_allocation
+            // Use u128 intermediate math to prevent overflow when pools exceed ~4 SOL
+            // (u64 overflows at ~1.8×10^19, but lamport products easily reach that)
             let winnings = if first_pool > 0 {
-                place_allocation
-                    .checked_mul(winning_deployed)
+                (place_allocation as u128)
+                    .checked_mul(winning_deployed as u128)
                     .ok_or(RumbleError::MathOverflow)?
-                    .checked_div(first_pool)
-                    .ok_or(RumbleError::MathOverflow)?
+                    .checked_div(first_pool as u128)
+                    .ok_or(RumbleError::MathOverflow)? as u64
             } else {
                 0
             };
